Feat: added component confirmation dialog Used for remove all tutorials and added button for publish tutorials

+++ Angular11/frontend/src/app/components/tutorials-list/tutorials-list.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
-import { NotificationMsg, NotificationType } from 'src/app/models/notification';
+import { NotificationType } from 'src/app/models/notification';
 import { Tutorial } from 'src/app/models/tutorial.model';
 import { NotificationService } from 'src/app/services/notification/notification.service';
 import { TutorialService } from 'src/app/services/tutorial/tutorial.service';
+import { MatDialogComponent, ConfirmDialogData } from '../mat-dialog/mat-dialog.component';
 @Component({
   selector: 'app-tutorials-list',
   templateUrl: './tutorials-list.component.html',
@@ -15,7 +16,9 @@ export class TutorialsListComponent implements OnInit{
   currentIndex = -1;
   titleField = new FormControl();
   constructor(private tutorialService: TutorialService,
-    private notificationService: NotificationService) { }
+    private notificationService: NotificationService, 
+    private dialog: MatDialogComponent) { }
+
   ngOnInit(): void {
     this.retrieveTutorials();
   }
@@ -39,16 +42,25 @@ export class TutorialsListComponent implements OnInit{
     this.currentIndex = index;
   }
   removeAllTutorials(): void {
-    this.tutorialService.deleteAll()
-      .subscribe(
-        response => {
-          console.log(response);
-          this.refreshList();
-          this.notificationService.sendNotification("All tutorials were deleted", NotificationType.info);
-        },
-        error => {
-          console.log(error);
-        });
+    this.dialog.openDialog(new ConfirmDialogData("Confirm your selection", "Delete all tutorials?")).subscribe(
+      selection => {
+        if(selection){
+          this.tutorialService.deleteAll()
+            .subscribe(
+              response => {
+                console.log(response);
+                this.refreshList();
+                this.notificationService.sendNotification("All tutorials were deleted", NotificationType.info);
+              },
+              error => {
+                console.log(error);
+              });
+          }
+      },
+      error => {
+        console.log(error);
+      }
+    );      
   }
   searchTitle(): void {
     this.tutorialService.findByTitle(this.titleField.value)
@@ -62,5 +74,24 @@ export class TutorialsListComponent implements OnInit{
         error => {
           console.log(error);
         });
+  }
+
+  publishTutorial(status: boolean): void {
+    this.currentTutorial!.published = status;
+    this.tutorialService.update(this.currentTutorial!._id, {published: status}).subscribe(
+      response => {},
+      error => {
+        console.log(error);
+      });
+  }
+
+  editTutorial(): void {  
+    this.tutorialService.get(this.currentTutorial!._id).subscribe(
+      response => {
+        console.log(response);
+      },
+      error => {        
+        console.log(error);
+      });
   }
 }
